# swagger.v0.1.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.1",
    "title": "BookLists Service",
    "description": "CRUD for Book Lists"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "tags": [
    {
      "name": "recap",
      "description": "Book Lists Service"
    }
  ],
  "paths": {
    "/v0.1/book-lists/{type}/{date}": {
      "get": {
        "tags": [
          "book-lists"
        ],
        "summary": "Get book-list by type, date",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "description": "Book list type (e.g. staff-picks)",
            "required": true,
            "type": "string"
          },
          {
            "name": "date",
            "in": "path",
            "description": "Book list date string (e.g. 2017-01)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Book List as json",
            "schema": {
              "$ref": "#/definitions/BookList"
            }
          },
          "500": {
            "description": "Error message in JSON format"
          }
        },
        "security": [
          {
            "api_auth": [
              "read:book-list"
            ]
          }
        ]
      }
    },
    "/v0.1/book-lists": {
      "get": {
        "tags": [
          "book-lists"
        ],
        "summary": "Get all book-lists",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Book list type (e.g. staff-picks)",
            "required": false,
            "type": "string"
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Number of results to return. Default 20",
            "required": false,
            "type": "integer"
          },
          {
            "name": "from_slug",
            "in": "query",
            "description": "For paginating, set from_slug to the last slug seen to retrieve all lists after that point.",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Book List stubs as json",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/BookList"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error message in JSON format"
          }
        },
        "security": [
          {
            "api_auth": [
              "read:book-list"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "book-lists"
        ],
        "summary": "Post new/updated book-list",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "booklist",
            "in": "body",
            "description": "Book list as JSON",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BookList"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book List stub as json",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/BookList"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error message in JSON format"
          }
        },
        "security": [
          {
            "api_auth": [
              "write:book-list"
            ]
          }
        ]
      }
    }
  },
  "definitions": {
    "BookList": {
      "title": "BookList",
      "type": "object",
      "properties": {
        
        "date": {
          "type": "string",
          "example": "2017",
          "description": "String representing the date of publication, with either year or month specificity"
        },
        "heroImageUrl": {
          "type": "string",
          "description": "URL of hero image",
          "required": false
        },
        "picks": {
          "type": "array",
          "required": false,
          "items": {
            "$ref": "#/definitions/BookListPick"
          }
        },
        "slug": {
          "type": "string",
          "description": "Identifier for list unique across all list types. Conventionally composed of type and date",
          "example": "staff-picks/2017-01"
        },
        "title": {
          "description": "User facing name of list",
          "type": "string",
          "example": "Best Books for Teens 2017",
          "required": false
        },
        "type": {
          "type": "string",
          "enum": [
            "staff-picks",
            "kids",
            "teens"
          ]
        }
      }
    },
    "BookListPick": {
      "title": "Pick",
      "type": "object",
      "properties": {
        "ageGroup": {
          "type": "string",
          "description": "Named age group? (e.g. \"Adult\") (Appears in Staff Picks)"
        },
        "book": {
          "$ref": "#/definitions/BookListBook"
        },
        "reviews": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BookListReview"
          }
        },
        "tags": {
          "type": "array",
          "description": "0-many tags from controlled vocabulary",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "BookListReview": {
      "title": "BookListReview",
      "type": "object",
      "properties": {
        "reviewerName": {
          "type": "string",
          "description": "Name of reviewer"
        },
        "reviewerLocation": {
          "type": "object",
          "description": "Location of reviewer",
          "properties": {
            "@id": { 
              "type": "string",
              "description": "Org id"
            },
            "prefLabel": {
              "type": "string",
              "description": "Display label for location",
              "example": "Volunteers Office"
            }
          }
        },
        "text": {
          "type": "string",
          "description": "Text of review"
        }
      }
    },
    "BookListBook": {
      "title": "BookListBook",
      "type": "object",
      "properties": {
        "author": {
          "type": "string",
          "description": "Name of author"
        },
        "catalogUrl": {
          "type": "string",
          "description": "Catalog link"
        },
        "ebookUrl": {
          "type": "string",
          "description": "E-book link"
        },
        "illustrator": {
          "type": "string",
          "description": "Name of illustrator (opt)"
        },
        "imageUrl": {
          "type": "string",
          "description": "Content Cafe link"
        },
        "isbn": {
          "type": "string",
          "description": "Not sure where used. Appears in Staff Picks"
        },
        "overdriveId": {
          "type": "string",
          "description": "Overdrive ID"
        },
        "publisher": {
          "type": "string",
          "description": "Name of publisher (opt)"
        },
        "title": {
          "type": "string",
          "description": "title of book"
        },
        "translator": {
          "type": "string",
          "description": "Name of translator (opt)"
        }
      }
    }
  }
}